[erlang] Don't build with wx or megaco
None of our applications use megaco or wx directly. wx is occasionally
useful in debugging situations, but in these cases it is most likely
that we would use a local erlang installation with wx enabled and
connect it to the remote host.  These apps are the two largest apps in
the erlang installation.

Signed-off-by: Steven Danna <steve@chef.io>