Add berkshelf-no-depselector software definition
The berkshelf-no-depselector software definition installs berkshelf from
source but

- without a dep-selector-libgecode dependency, and
- excluding many bundler groups

Berks only requires dep-selector-libgecode for cases where the user
desires resolution results *exactly* like the chef-server would
provide. Most of our users of berks is to easily manage very small set
of cookbooks that omnibus packages use for configuration.

This resulting berks install is smaller and faster to install than the
standard berkshelf software definition and eliminates most of the cost
of using berks.

Installing from gems would be best, but will require someone to do the
work to collect all of the transitive licensing information.

We can likely collapse these if someone audits our use of berkshelf
across that packages and verifies that a no-dep-selector install will
work for all of them (we pin to some old versions in a number of
projects).

Signed-off-by: Steven Danna <steve@chef.io>