Use a github mirror of GNU config
Savannah has been unavailable multiple times this year. We don't
typically need frequent updates to the repository.  I've pushed a copy
of it to github, a service which tends to be more reliable for us.

If we want to use this long term, we should find a way to update our
mirror periodically.

Signed-off-by: Steven Danna <steve@chef.io>